requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
sentence-transformers==2.5.1
selenium==4.18.1
webdriver-manager==4.0.1
pillow==10.2.0
python-multipart==0.0.9
numpy==1.26.4
scikit-learn==1.4.1.post1
torch==2.2.1
transformers==4.38.2
beautifulsoup4==4.12.3 